requirements.txt
smolagents
textual
python-dotenv
lancedb>=0.5.0
numpy>=1.21.0
requests>=2.28.0
pyarrow>=10.0.0
